fix(plugin/server): compat & concurrency defects (ultrareview wave 5)
- H2  delete_element / batch_rename / change_element_type: int[] -> long[] so
      64-bit Revit 2024+ element IDs are no longer silently truncated.
- H10 CodeDomExecutor: use version-guarded GetElementIdValue instead of
      elem.Id.IntegerValue (truncates large IDs on R2024/net48).
- H5  RevitThreadDispatcher: release the global lock before WaitForCompletion so
      one slow op no longer serializes all tool calls; clear prepared state on
      timeout so the next request isn't refused.
- H27 CortexRouter._disabledTools: volatile copy-on-write HashSet swapped
      atomically (Immutable* is unavailable on net48), fixing the UI-thread-write
      vs socket-thread-read race.
- H1  UpdateChecker.LaunchInstaller now returns bool; both callers only schedule
      Revit shutdown when the installer actually launched (UAC-deny safe).
- H6  ProtocolHandlerRegistrar: decode params with [uri]::UnescapeDataString
      instead of System.Web before its assembly was loaded.
- H15 PowerBiExportWindow.SetStatus marshals to the Dispatcher when called off
      the UI thread (post-await continuations).

Builds green on R25 (net8), R24 (net48), and the MCP server; 351 tests pass.

+++ src/RevitCortex.Server/Tools/CreationTools.cs
@@ -65,7 +65,7 @@ public static class CreationTools
     [McpServerTool(Name = "change_element_type"), Description("Change the type of one or more elements to a target type specified by ID or name.")]
     public static async Task<string> ChangeElementType(
         RevitConnectionManager revit,
-        [Description("Element IDs to change")] int[] elementIds,
+        [Description("Element IDs to change")] long[] elementIds,
         [Description("Target type element ID")] long? targetTypeId = null,
         [Description("Target type name")] string? targetTypeName = null,
         CancellationToken ct = default)
